added header comp&moved navbar into seperate dir
