+++ NextGen-BM-FE/src/components/Header/HeaderStyles.tsx
@@ -1,0 +1,11 @@
+import { SxProps } from "@mui/material";
+
+export const headerStyles:{[key:string]:SxProps}={
+    appBarStyles:{
+        position: 'fixed',
+        width: '100%',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'space-between'
+    }
+};
