+++ NextGen-BM-FE/src/components/Navbar/NavbarStyles.ts
@@ -1,8 +1,10 @@
-export const navbarStyles={
+import { SxProps } from "@mui/material";
+
+export const navbarStyles:{[key:string]:SxProps}={
     drawerStyles:{
         boxSizing:'border-box',
         width: '30%'
-      },
+    },
     paperProps:{
         width:'30%'
     }
