+++ NextGen-BM-FE/src/layouts/MainLayout.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import Navbar from "../components/Navbar";
+import Navbar from "../components/Navbar/Navbar.tsx";
 import { Outlet } from "react-router-dom";
 import { Box } from "@mui/material";
 import { Header } from "../components/Header/Header.tsx";
